[Refactoring ⚙️] 로딩스피너 애니메이션 관련 로직 useLoading 훅으로 분리
